DEMO: collclean_lib.py
This is the code currently running at DESY to semi-standardise collaboration names
and find the corresponding experiment.
Some parts are du to 'features' in the metadata from arXiv and might be obsolete.
E.g author names in the collaboration field.

It might be better to take information about e.g. capitalisation from the expriments database.

Signed-off-by: Kirsten Sachs <kirsten.sachs@desy.de>